Update UserMapper.java
s

+++ myweb/src/main/java/org/boli/myweb/mapper/UserMapper.java
@@ -6,25 +6,25 @@ import org.boli.myweb.entity.User;
 import org.boli.myweb.entity.UserExample;
 
 public interface UserMapper {
-    long countByExample(UserExample example);
+	long countByExample(UserExample example);
 
-    int deleteByExample(UserExample example);
+	int deleteByExample(UserExample example);
 
-    int deleteByPrimaryKey(Integer id);
+	int deleteByPrimaryKey(Long id);
 
-    int insert(User record);
+	int insert(User record);
 
-    int insertSelective(User record);
+	int insertSelective(User record);
 
-    List<User> selectByExample(UserExample example);
+	List<User> selectByExample(UserExample example);
 
-    User selectByPrimaryKey(Integer id);
+	User selectByPrimaryKey(Long id);
 
-    int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
+	int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
 
-    int updateByExample(@Param("record") User record, @Param("example") UserExample example);
+	int updateByExample(@Param("record") User record, @Param("example") UserExample example);
 
-    int updateByPrimaryKeySelective(User record);
+	int updateByPrimaryKeySelective(User record);
 
-    int updateByPrimaryKey(User record);
+	int updateByPrimaryKey(User record);
 }
